Move statuscodes to core
