+++ parcels/interaction/interactionkernel.py
@@ -7,7 +7,7 @@ import numpy as np
 from parcels._compat import MPI
 from parcels._core.field import VectorField
 from parcels.kernel import BaseKernel
-from parcels.statuscodes import StatusCode
+from parcels._core.statuscodes import StatusCode
 
 __all__ = ["InteractionKernel"]
