+++ parcels/kernels/advection.py
@@ -4,7 +4,7 @@ import math
 
 import numpy as np
 
-from parcels.statuscodes import StatusCode
+from parcels._core.statuscodes import StatusCode
 
 __all__ = [
     "AdvectionAnalytical",
