+++ tests/test_particlefile.py
@@ -14,7 +14,7 @@ from parcels._datasets.structured.generic import datasets
 from parcels.kernels import AdvectionRK4
 from parcels._core.particle import Particle, create_particle_data, get_default_particle
 from parcels import ParticleFile
-from parcels.statuscodes import StatusCode
+from parcels import StatusCode
 from parcels import XGrid
 from tests.common_kernels import DoNothing
